fix(act): u8/i64 arithmetic overflow in decomposition + msg_id
1. sia_loop decompose_task: (10 - i as u8) underflowed u8 at the 11th
   sentence (debug panic / release wrap to 255). Now 10u8.saturating_sub.
2. per_agent planner: 10u8.saturating_sub(i as u8 * 2) overflowed at the
   128th segment (i*2 > 255). Now i.saturating_mul(2).
3. gram session next_msg_id: (now as i64) << 32 overflows after 2038.
   Shift in u64, cast to i64 — preserves MTProto layout.

Tests: test_decompose_many_sentences_no_priority_underflow (20 sentences),
test_planner_many_steps_no_priority_overflow (200 segments).

+++ neotrix-core/src/neotrix/l1_body_impl/nt_act_gram/server/session.rs
@@ -27,9 +27,12 @@ impl Session {
             .duration_since(std::time::UNIX_EPOCH)
             .unwrap_or_default()
             .as_secs();
-        let base = (now as i64) << 32;
+        // Seconds in the high 32 bits, per-turn counter in the low 32 bits
+        // (MTProto msg_id layout). Shift in u64 so (now >= 2^31, i.e. after
+        // 2038) cannot overflow i64 in debug builds.
+        let base = (now as u64) << 32;
         let prev = self.msg_id.fetch_add(1, Ordering::SeqCst);
-        base | (prev as i64 & 0xFFFFFFFF)
+        (base | (prev as u64 & 0xFFFFFFFF)) as i64
     }
 
     pub fn is_authenticated(&self) -> bool {
